+++ pkg/client/cli/session.go
@@ -81,7 +81,7 @@ func (si *sessionInfo) connect(cmd *cobra.Command, args []string) error {
 		return si.withConnector(true, func(_ *connectorState) error { return nil })
 	}
 	return si.withConnector(false, func(cs *connectorState) error {
-		return start(args[0], args[1:], true, cmd.InOrStdin(), cmd.OutOrStdout(), cmd.ErrOrStderr())
+		return start(cmd.Context(), args[0], args[1:], true, cmd.InOrStdin(), cmd.OutOrStdout(), cmd.ErrOrStderr())
 	})
 }